factored out scale resolution into utility method
